requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
fasteners==0.16.3
flatbuffers==1.12
future==0.18.2
gast==0.2.2
gin-config==0.4.0
glfw==2.1.0
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==3.10.0
kiwisolver==1.3.1
labmaze==1.0.5
lockfile==0.12.2
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.4
mujoco-py==2.0.2.13
networkx==2.5.1
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
Pillow==7.2.0
Pillow-SIMD==7.0.0.post3
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.0
pycparser==2.20
pyglet==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.5.4
seaborn==0.11.1
six==1.16.0
sortedcontainers==2.4.0
tensorboard==2.0.2
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2020.9.3
torch==1.8.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1